+++ Application/Common/Models/LanguageDto.cs
@@ -2,8 +2,10 @@
 {
     public class LanguageDto
     {
-        public int Id { get; set; }
+        public short Id { get; set; }
+
         public string? Title { get; set; }
+
         public string? LanguageCode { get; set; }
 
     }
