+++ Application/Common/Models/RoleDto.cs
@@ -8,9 +8,8 @@ namespace Application.Common.Models
 {
     public class RoleDto
     {
-        public int Id { get; set; }
-
+        public short Id { get; set; }
         public string? Title { get; set; }
-
+   
     }
 }
