+++ Application/Common/Models/UserDto.cs
@@ -3,7 +3,6 @@
     public class UserDto
     {
         public int Id { get; set; }
-
         public string? Name { get; set; }
 
         public string? Surname { get; set; }
@@ -14,6 +13,6 @@
 
         public string? Password { get; set; }
 
-        public int RoleId { get; set; }
+        public short RoleId { get; set; }
     }
 }
