+++ Domain/Entities/PostHashtag.cs
@@ -5,9 +5,9 @@ namespace Domain.Entities
 {
     public sealed class PostHashtag : BaseAuditableEntity
     {
-        public int NewsId { get; set; }
+        public long NewsId { get; set; }
 
-        public int HashtagId { get; set; }
+        public long HashtagId { get; set; }
 
         public PostTranslation? PostTranslation { get; set; }
